requirements.txt
django
django-crispy-forms==1.14.0
django-cleanup==5.2.0
pillow
django-ckeditor
channels
django-allauth
python-dotenv
django-braces
# https://docs.djangoproject.com/en/4.1/ref/contrib/humanize/
humanize
django-debug-toolbar
# https://jupyter.org/install
notebook
# pip install django-extensions
tokens